pyproject.toml
[tool.poetry]
name = "samplesdb"
version = "0.1.0"
description = ""
authors = ["Paul CHAUMEIL <paul@stackadoc.com>", "Arthur RENAUD <arthur@stackadoc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.24"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
